chore: Update Dockerfile to include pnpm, corepack, and additional dependencies
This commit updates the Dockerfile to include the installation of pnpm and corepack, as well as additional dependencies required for the application. The unnecessary user and working directory configurations have been removed. By including these dependencies in the Dockerfile, we ensure that they are available during the build process and when the application is deployed. This change improves the efficiency and maintainability of the Dockerfile.