chore: Update Dockerfile to include pnpm and corepack
This commit updates the Dockerfile to include the installation of pnpm and corepack. These dependencies are necessary for the application to run properly. By including them in the Dockerfile, we ensure that they are available during the build process and when the application is deployed. This change improves the efficiency and maintainability of the Dockerfile.