chore: Update deploy-preview.yml to pull from the 'preview' branch
This commit updates the deploy-preview.yml file to change the branch from which the code is pulled during deployment. Instead of pulling from the 'main' branch, it now pulls from the 'preview' branch. This change ensures that the latest changes in the 'preview' branch are deployed for testing and review purposes.