chore: Update prisma schema to include UserPreferences and UserComms models
This commit adds the UserPreferences and UserComms models to the prisma schema. These models represent the preferences and communication channels associated with a user. By including them in the schema, we ensure that the necessary tables are created in the database and provide a structured way to store and retrieve user preferences and communication information.