```text chore: Update npm dependencies and add Puppeteer plugins
This commit updates the npm dependencies in the package.json file to include the latest stable versions of Puppeteer and other related packages. It also adds Puppeteer plugins, such as puppeteer-extra-plugin-devtools, puppeteer-extra-plugin-stealth, and puppeteer-jquery. These updates and additions improve the functionality and performance of the application when interacting with web pages and automate tasks.